Renamed kunta-api-embeds -> embeds.php in management
